+++ hashtags/hashtags/settings.py
@@ -51,7 +51,7 @@ INSTALLED_APPS = (
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'hashtags',
-    'posts',
+    'tags',
 )
 
 MIDDLEWARE_CLASSES = (
