+++ hashtags/hashtags/views.py
@@ -2,8 +2,8 @@
 from django.http import HttpResponse
 from django.shortcuts import render
 
-from posts.models import Post, Tag
-from posts.forms import SearchForm
+from tags.models import Post, Tag
+from tags.forms import SearchForm
 
 import json
